use plain js for simple turn background page
