+++ src/samples/simple-turn-chromeapp/manifest.json
@@ -6,7 +6,7 @@
   "app": {
     "background": {
       "scripts": [
-        "background.core-env.static.js"
+        "background.js"
       ]
     }
   },
